fix: resolve type casting error in ResponsiveMacroIndicators landscape mode
Fixed type casting issue when rotating to landscape orientation:
- Changed clamp(40, 60) to clamp(40.0, 60.0) to ensure double return type
- Removed unnecessary 'as double' cast that was failing with int return
- clamp() with double arguments now returns double without explicit casting

This fixes the error: "type 'int' is not a subtype of type 'double' in type cast"
when MacroIndicators component was displayed in landscape mode.

+++ lib/core/widgets/charts/macro_indicators.dart
@@ -565,7 +565,7 @@ class ResponsiveMacroIndicators extends StatelessWidget {
       builder: (context, constraints) {
         // Adjust indicator radius based on screen width
         final availableWidth = constraints.maxWidth;
-        final indicatorRadius = ((availableWidth / 6).clamp(40, 60) as double);
+        final indicatorRadius = (availableWidth / 6).clamp(40.0, 60.0);
 
         return MacroIndicators(
           macroData: macroData,
